Implemented a bogus,semirandom version of most constrainted variable using the number of neighbors
