+++ FreeFlow/Outputs/55.txt
@@ -4,5 +4,5 @@ BRYOO
 BROOG
 BBGGG
 
-Time Taken 0.046874046325683594 seconds
-Number of Assignments: 29
+Time Taken 0.03355693817138672 seconds
+Number of Assignments: 20
