+++ FreeFlow/Outputs/77.txt
@@ -1,10 +1,10 @@
-GGGGGGG
-GBBYYYY
-GGBYRRY
-OGBYRYY
-OGRRRYO
-OYYYYYO
-OOOOOOO
+GGGOOOO
+GBGGGYO
+GBBBRYO
+GYYYRYO
+GYRRRYO
+GYRYYYO
+GYYYOOO
 
-Time Taken 0.07361102104187012 seconds
-Number of Assignments: 148
+Time Taken 0.7091519832611084 seconds
+Number of Assignments: 133
