+++ FreeFlow/Outputs/88.txt
@@ -1,11 +1,11 @@
+YYYRRRGG
+YBYPPRRG
+YBOOPGRG
+YBOPPGGG
+YBOOOOYY
+YBBBBOQY
+YQQQQQQY
 YYYYYYYY
-YBBBBBQY
-YYOOOBQY
-RPOPOBQY
-RPPPOBQY
-RRGGOOQY
-GRRGYQQY
-GGGGYYYY
 
-Time Taken 0.3948960304260254 seconds
-Number of Assignments: 405
+Time Taken 1014.2205040454865 seconds
+Number of Assignments: 93148
