+++ FreeFlow/Outputs/991.txt
@@ -1,12 +1,12 @@
-DDDDGGGGG
-BBBBGKKKG
-BORRKKYKG
-BOQRKPYKG
-OOQRKPYKG
-KRQRKPYKG
-KRQRKPYKG
-KRRRKPPPG
-KKKKKGGGG
+DBBBOKKKK
+DBOOORRRK
+DBRQQQQRK
+DBRRRRRRK
+GGKKKKKKK
+GKKPPPPPG
+GKYYYYYPG
+GKKKKKKPG
+GGGGGGGGG
 
-Time Taken 2.3646299839019775 seconds
-Number of Assignments: 1497
+Time Taken 112.91026401519775 seconds
+Number of Assignments: 7397
