+++ FreeFlow/smart_solver.py
@@ -1,6 +1,7 @@
 # In This file only write the main function. Define functions in either readInput or CSPSolving
 from readInput import *
 from CSPsmart import *
+from heuristics import *
 import time
 import random
 
@@ -59,42 +60,42 @@ def smart_solver(solve_dict,height,width,color_set,initial_points):
 # same bit set (or with neither bit set) for any allowed colour.]
 
 if __name__ == "__main__":
-    games = get_list_of_smaller_files()
-    for gameboard in games:
-        #gameboard=games[-1]
-        name = get_name(gameboard)
-        useful_array_board = input_to_array(gameboard)
-        height, width = puzzleDetails(useful_array_board)
-        print("height:", height, "width:", width)
-        for row in useful_array_board:
-            print(row)
+    games = get_list_of_bigger_files()
+    #for gameboard in games:
+    gameboard=games[0]
+    name = get_name(gameboard)
+    useful_array_board = input_to_array(gameboard)
+    height, width = puzzleDetails(useful_array_board)
+    print("height:", height, "width:", width)
+    for row in useful_array_board:
+        print(row)
 
-        print('\n')
+    print('\n')
 
-        color_set, solve_dict = generateColorSet_Dict(useful_array_board)
+    color_set, solve_dict = generateColorSet_Dict(useful_array_board)
 
-        print("colorSet:", color_set, "colorDict:", solve_dict)
-        initial_points = solve_dict.copy()
-        #THESE COMMENTS ARE IMPORTANT EVENTUALLY WE'RE SUPPOSED TO HAVE AVERAGES so I figured 10 is good
-        #average_time=0
-        #average_assignments=0
-        #for i in range(0,10):
-        solved_maze_input=solve_dict.copy()
-        start=time.time()
-        assignments=0
-        solved_maze=smart_solver(solved_maze_input,height,width,color_set,initial_points)
-        end=time.time()
-        #average_time += (end-start)
-        #average_assignments +=assignments
-        #average_assignments=average_assignments/10
-        #average_time=average_time/10
+    print("colorSet:", color_set, "colorDict:", solve_dict)
+    initial_points = solve_dict.copy()
+    #THESE COMMENTS ARE IMPORTANT EVENTUALLY WE'RE SUPPOSED TO HAVE AVERAGES so I figured 10 is good
+    #average_time=0
+    #average_assignments=0
+    #for i in range(0,10):
+    solved_maze_input=solve_dict.copy()
+    start=time.time()
+    assignments=0
+    solved_maze=smart_solver(solved_maze_input,height,width,color_set,initial_points)
+    end=time.time()
+    #average_time += (end-start)
+    #average_assignments +=assignments
+    #average_assignments=average_assignments/10
+    #average_time=average_time/10
 
-        print('\n')
-        print(end-start)
-        print(assignments)
-        print(solved_maze)
-        print('\n')
-        print_free_flow(solved_maze,height,width)
-        filename = 'Outputs/%s.txt' % name
-        print_free_flow_file(solved_maze, height, width,str(end-start),str(assignments),filename)
-        print('______________________________________________')
+    print('\n')
+    print(end-start)
+    print(assignments)
+    print(solved_maze)
+    print('\n')
+    print_free_flow(solved_maze,height,width)
+    filename = 'Outputs/%s.txt' % name
+    print_free_flow_file(solved_maze, height, width,str(end-start),str(assignments),filename)
+    print('______________________________________________')
